Fixed broken mint/hunt calculator
Okay, this SHOULD be working, but I'm not 100% certain that I have the
maths right... the original variables were poorly named and there wasn't
any documentation, no comments, I basically had to try and reverse
engineer the original calculator and then reimplement it properly.

Also, this slightly expands the size of the options panel.