Reorganisation, should load options on injection
It should now load (and apply) options like the time reformatter as soon
as the script loads now... also, this _might_ fix a possible
(unreported, so it might just be my machine :v) memory issue that had
the page failing to render or respond after extended periods. If not...
I'll keep bashing it until it does.