Bugfixed several script breakers
Reverted to the old method of shortTimeFormat, but kept days included.
The unicorn calculator now shows a message in the details section if you
don't have what you need for it to run, so it doesn't look broken. All
of the previous instances of `game.bld.getHyperbolicEffect` have been
migrated to `game.getHyperbolicEffect` now, which I assume is the same
function in a different place. If not, I'm gonna yell at the dev.

Shoutout to Quizer on the game's discord server for bringing this to my
attention, since my current game doesn't have Ziggurats yet.