Fixed the unicorn calculator
Previously, I hadn't touched this bit, so it was trying to do maths with
undefined values, which gave it NaN and broke everything. I've dug
through as much as I can understand at the moment and at the very least
I fixed the NaN and a subsequently discovered crashing bug - unguarded
division by a potentially-zero value led to a call stack overflow in the
game's code where it was never meant to handle Infinity as a value.

If anyone would like to go over the relevant maths, or honestly anything
else for that matter, I'd appreciate any help.