Allow disabling pollution, but for real
The game doesn't ACTUALLY disable pollution with the option for it. It
only hides the pollution info in the top bar. This is a simple-as-hell
hack to intercept pollution and forcibly lock it to zero. You can't
restore your previous pollution value if you should want to for some
reason, so don't use it unless you're willing to fully reset your
pollution level to zero.